Refactoring QuickOpen extensions, separating reused code into a new file.
